* Refactors webpack config from JS to TS
* Adds types for webpack and its dependencies

* Adds types for webpack plugins currently in use

* Updates terser-webpack-plugin from 4.0.2 to 4.2.3
