+++ package.json
@@ -11,6 +11,12 @@
 		"@babel/plugin-transform-runtime": "^7.11.5",
 		"@babel/preset-env": "^7.11.5",
 		"@babel/runtime": "^7.11.2",
+		"@types/mini-css-extract-plugin": "^0.9.0",
+		"@types/optimize-css-assets-webpack-plugin": "^5.0.1",
+		"@types/node": "^14.14.7",
+		"@types/terser-webpack-plugin": "^4.2.0",
+		"@types/webpack": "^4.41.25",
+		"@types/webpack-dev-server": "^3.11.1",
 		"autoprefixer": "^9.8.6",
 		"babel-loader": "^8.1.0",
 		"clean-webpack-plugin": "^3.0.0",
@@ -25,9 +31,10 @@
 		"svelte-check": "^1.0.32",
 		"svelte-loader-hot": "^0.3.1",
 		"svelte-preprocess": "^4.2.0",
-		"terser-webpack-plugin": "^4.0.0",
+		"terser-webpack-plugin": "^4.2.3",
 		"ts-loader": "^8.0.2",
-		"typescript": "^4.0.2",
+		"ts-node": "^9.0.0",
+		"typescript": "^4.0.5",
 		"webpack": "^4.44.1",
 		"webpack-cli": "^3.3.12",
 		"webpack-dev-server": "^3.11.0"
